@@ -1,37 +1,28 @@
-import React, { Component } from 'react';
-import { View, Button } from 'react-native';
-import PasswordInputText from "react-native-hide-show-password-input";
+// import packages
+import React, {useState} from 'react';
+import {View, Button} from 'react-native';
+import PasswordInputText from 'react-native-hide-show-password-input';
 
-export default class App extends Component {
+export const App = () => {
+  const [password, setPassword] = useState('');
+  let inputRef = null;
 
-  constructor(props) {
-    super(props);
-    this.state = {
-      password: '',
-    };
-  }
+  return (
+    <View style={{margin: 20}}>
+      <PasswordInputText
+        getRef={(input) => (inputRef = input)}
+        value={password}
+        onChangeText={(password) => setPassword(password)}
+      />
 
+      <Button
+        title="Clear"
+        onPress={() => {
+          if (inputRef) inputRef.clear();
+        }}
+      />
+    </View>
+  );
+};
 
-  render() {
-
-    const { password } = this.state;
-
-    return (
-      <View style={{ margin: 20 }}>
-
-        <PasswordInputText
-          getRef={input => this.input = input}
-          value={password}
-          onChangeText={(password) => this.setState({ password })}
-        />
-
-        <Button
-          title="Clear"
-          onPress={() => this.input.clear()} />
-          
-      </View>
-    );
-  }
-}
-
-
+export default App;
